# backend/app/openapi/api_docs.yaml
openapi: 3.0.0
info:
  title: Task Manager API
  description: API для управления задачами с аутентификацией пользователей
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        detail:
          type: string

paths:
  /api/auth/register:
    post:
      summary: Регистрация нового пользователя
      tags:
        - authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
        '400':
          description: Ошибка валидации или пользователь уже существует
        '429':
          description: Слишком много попыток регистрации

  /api/auth/login:
    post:
      summary: Аутентификация пользователя
      tags:
        - authentication
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
        '401':
          description: Неверные учетные данные
        '429':
          description: Слишком много попыток входа

  /api/tasks:
    get:
      summary: Получение списка задач
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Создание новой задачи
      tags:
        - tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '201':
          description: Задача создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /api/tasks/{task_id}:
    parameters:
      - in: path
        name: task_id
        required: true
        schema:
          type: integer
    get:
      summary: Получение задачи по ID
      tags:
        - tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Задача найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Задача не найдена
    put:
      summary: Обновление задачи
      tags:
        - tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '200':
          description: Задача обновлена
        '404':
          description: Задача не найдена
    delete:
      summary: Удаление задачи
      tags:
        - tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Задача удалена
        '404':
          description: Задача не найдена